# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Haife API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/prizes": {
      "post": {
        "tags": [
          "Prizes"
        ],
        "summary": "新增獎項",
        "description": "傳入獎項資訊並建立獎項",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "使用者權杖",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "獎項新增資訊",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "獎項名稱"
                },
                "price": {
                  "type": "string",
                  "example": "價格"
                },
                "time": {
                  "type": "string",
                  "example": "個數"
                },
                "quantity": {
                  "type": "string",
                  "example": "獎項數量"
                },
                "remainingQty": {
                  "type": "string",
                  "example": "剩餘數量"
                },
                "type": {
                  "type": "string",
                  "example": "獎項類型"
                },
                "isEnabled": {
                  "type": "string",
                  "example": "是否啟用"
                },
                "activity_uuid": {
                  "type": "string",
                  "example": "活動編號"
                }
              },
              "required": [
                "name",
                "price",
                "time",
                "quantity",
                "remainingQty",
                "type",
                "isEnabled",
                "activity_uuid"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "新增成功"
          },
          "400": {
            "description": "新增失敗"
          },
          "403": {
            "description": "使用者沒有角色頁面權限"
          }
        }
      }
    },
...
    "/api/draw": {
      "post": {
        "tags": [
          "Draw"
        ],
        "summary": "抽獎",
        "description": "進行抽獎",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "抽獎序號",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "抽獎者資訊",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "抽獎者姓名"
                },
                "phone": {
                  "type": "string",
                  "example": "聯絡電話"
                },
                "designer": {
                  "type": "string",
                  "example": "設計師"
                }
              },
              "required": [
                "userName",
                "phone",
                "designer"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "抽獎成功"
          },
          "400": {
            "description": "抽獎失敗"
          }
        }
      }
    },
    "/api/draw/verify": {
      "post": {
        "tags": [
          "Draw"
        ],
        "summary": "查核抽獎序號",
        "description": "查核抽獎序號",
        "parameters": [
          {
            "name": "code",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "example": "any"
                }
              }
            },
            "required": true,
            "description": "抽獎序號"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          }
        }
      }
    },
    "/api/draw/prizes": {
      "get": {
        "tags": [
          "Draw"
        ],
        "summary": "取得所有獎項",
        "description": "列出所有獎項",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "抽獎序號",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          }
        }
      }
    },
    "/api/draw/winner": {
      "get": {
        "tags": [
          "Draw"
        ],
        "summary": "取得中獎獎項",
        "description": "取得中獎獎項",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "抽獎序號",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功"
          },
          "400": {
            "description": "查詢失敗"
          }
        }
      }
    }
  }
}